requirements.txt
# ============================================================
# Requirements for chart-vit-xai
# Tested with Python 3.10.18
# ============================================================

# ------------------------------------------------------------
# Deep Learning
# Install PyTorch separately based on your CUDA version:
#   https://pytorch.org/get-started/locally/
# Tested: torch==2.8.0+cu128, torchvision==0.23.0+cu128 (CUDA 12.8)
# ------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0

# ------------------------------------------------------------
# Data Processing
# ------------------------------------------------------------
numpy==2.1.2
pandas==2.3.3
h5py==3.15.1
scikit-learn==1.7.2
scipy==1.15.3

# ------------------------------------------------------------
# Image Processing
# ------------------------------------------------------------
opencv-python==4.12.0.88
Pillow==11.0.0

# ------------------------------------------------------------
# Statistics (Backtest & Performance Metrics)
# ------------------------------------------------------------
statsmodels==0.14.6

# ------------------------------------------------------------
# Utilities
# ------------------------------------------------------------
tqdm==4.67.1
joblib==1.5.2
matplotlib==3.10.8

# ------------------------------------------------------------
# Experiment Tracking
# ------------------------------------------------------------
wandb==0.22.2

# ------------------------------------------------------------
# Jupyter (for notebook files)
# ------------------------------------------------------------
ipykernel==6.30.1
jupyter_client==8.6.3